requirements.txt
absl-py==2.1.0
arviz==0.18.0
asttokens==2.4.1
cachetools==5.3.2
certifi==2024.2.2
cftime==1.6.3
cloudpickle==3.0.0
cons==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
etuples==0.3.9
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
h5netcdf==1.3.0
h5py==3.11.0
ipdb==0.13.13
ipython==8.24.0
jedi==0.19.1
kiwisolver==1.4.5
logical-unification==0.4.6
markdown-it-py==3.0.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
miniKanren==1.0.3
multipledispatch==1.0.0
netCDF4==1.6.5
numpy==1.26.4
olefile==0.47
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pprintpp==0.4.0
prompt_toolkit==3.0.43
ptyprocess==0.7.0
pure_eval==0.2.2
Pygments==2.17.2
pymc==5.15.0
pyparsing==3.1.2
pytensor==2.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
rich==13.7.1
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
stack_data==0.6.3
toml==0.10.2
toolz==0.12.1
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
xarray==2024.2.0
xarray-einstats==0.7.0
